Update dependencies
This includes switching to aioredis 2.0, which required changes to
the Redis code and unfortunately broke mockaioredis.  Change how a
mock Redis is injected to move the code out of the main application
and disable the mock Redis tests for the time being.  This means that
only the Docker test that uses a real Redis server can be used for
the time being.

Also adjust for changes in the latest Safir.