Update conditional workflow logic
Enable CI workflow on pull requests, and suppress the CI workflow
on pushes to branches that should always have a pull request to
avoid duplicate work.  Redo the conditional on the build job to
run on pull requests from ticket branches.  Merge the docs job
into the main CI workflow and run it unconditionally, but only try
to upload the documentation for non-PR pushes or pull requests from
a ticket branch.