Stop using the FastAPI Docker image
Follow the new upstream recommendations and the new FastAPI project
template and build our own Docker image with a start script that
initializes the database and then starts a single process with
uvicorn.  Scaling in Kubernetes is better-handled by running multiple
pods rather than multiple processes in a pod.

Take advantage of this change to update the Docker base image to
Debian bullseye and the Python version to 3.9.8.  Require Python 3.9
or later and stop testing with Python 3.8.