Move Selenium code out of string template
There is a mechanism in uvicorn to start an app using a factory
function, so all of the Selenium code can be moved out of a string
template and put directly in tests.support.selenium, where it can
be linted and type-checked.

Use an environment variable instead of templating to communicate
the path to which to write the auth token to the Selenium testing
app.