Drop support for non-PostgreSQL databases
As with the support for mock Redis, the support for running with a
SQLite database is not currently enabled and not tested.  Dropping
support allows some simplification, and we don't intend to deploy
with SQLite.  The only cost is that tests are now forced to take
longer since testing with a real PostgreSQL database is quite a bit
slower, but we were already paying that cost.