Refactor token caching
Rather than making caching independent of creating new internal
and notebook tokens, have the cache manage the token creation and
create new tokens under a per-user asyncio lock.  This will ensure
that token creation is done only once and then cached and all
subsequent requests are served from that cache.

In the previous logic, all requests would fall through the cache,
and then all requests would have to query the database to
reconstruct the token, which in turn relies heavily on database
synchronization to ensure multiple tokens are not generated.