Further refactor token cache
Creating the token cache with locks at module import time won't
work properly with the test suite when making calls to the underlying
app because the locks will be created with a separate loop than the
app uses.  Instead, add a dependency that manages the token cache
data structure with locks, and create it lazily when it's first
needed.