Avoid charset_normalize warning
Parsing the error return from the OIDC server when given a bad
client ID was for some reason triggering a warning in
charset_normalize.  This is probably some sort of bug in it or
httpx, but the return status is JSON, so work around the problem
by parsing it as such.