Finish eliminating SetupTest
Move the remaining utility functions into other modules and create
a new factory fixture, which replaces the last use of SetupTest.

Add a reconfigure method to ComponentFactory to update it with a
new configuration, required by many of the tests.

Properly wrap database operations in tests with a transaction, now
that the code is designed for one transaction per request and doesn't
do transaction management inside the services.