Convert to kubernetes_asyncio
Use the asyncio Kubernetes library, and propagate the async
status through the storage and service layer.  Convert the watch
queue to asyncio.Queue accordingly.

This allows us to throw regular exceptions from the watch task,
since it's now a normal asyncio task, and gets rid of the custom
thread pool stuff we had to do to background the watch thread.