Clean up documentation build
Add nitpick_ignore for the autogenerated references that will
never resolve.  Add some missing modules to api.rst.  Fix lots of
other issues uncovered now that the documentation warnings aren't
full of noise.

Stop using TYPE_CHECKING for typing-only imports in the modules
that are documented.  This speeds up imports but it makes Sphinx
fail to resolve references (I tried a workaround without any luck),
and it's extra complexity to manage when maintaining the code.  We
don't care about fast startup for Gafaelfawr, so this is not really
buying us anything.

Bump the copyright year for the package.

Re-enable inheritance diagrams in the Sphinx API documentation.

Make warnings fatal in the documentation build to prevent
regressions.