Fix tests for latest pytest-asyncio
The new version of pytest-asyncio functions somewhat differently,
and the tests were not explicit about which fixtures were async
and which were sync, causing the cli tests to break because of a
conflict of event loops.

Fix this by switching to strict mode and being explicit about what
is async and what is sync, and then avoid using a necessarily-async
fixture for the necessarily-sync cli tests.