Switch to the Safir database layer
Remove the local database session setup functions and the
database session dependency and use the ones from Safir instead.
This switches Gafaelfawr from using a fresh AsyncSession for
each request to using an async_scoped_session, since that's the
approach used by vo-cutouts.

Safir now handles insertion of the database password into the
URL, so remove the code in configuration parsing that was doing
this and the corresponding test.

Having the Safir db_session_dependency open transactions creates
strange problems with FastAPI 0.74 or later where the commit of
the transaction after the yield in the db_session_dependency is
immediately cancelled with an asyncio.CancelledError.  This may be
some bug in asyncpg, sqlalchemy, or FastAPI, but explicitly
managing transactions isn't too hard and allows us to be consistent
with vo-cutouts.

Do the transaction management in the handlers, which avoids having
to introduce database sessions to the service layer and allows
encapsulating multiple database calls in the same transaction.