Refactor LDAP layer and OIDC token parsing
Move the LDAP code into an LDAPStorage object and out of the
OIDCProvider class.  We may eventually want to manage it with a
dependency, or allow it to be used with GitHub authentication,
and this makes the code easier to read.

Move parsing of groups from a token into TokenVerifier to keep
all upstream JWT parsing together.  Move UID extraction into a
separate method and remove UID as a top-level attribute of an
OIDCVerifiedToken, since it's now optional and we may never want
to look at those attributes of a token if LDAP support is
configured.