Use a single RSA key pair for all tests
Creating RSA key pairs takes a surprising length of time if one
creates a new one for every test or change to the settings file.
Create a single one during module import in the test support code,
and thus a single one per test run, and reuse it for every
settings file.