Use a single engine for all non-app tests
Create a single session-scoped database engine and reuse it for
all database operations other than ones performed by the app.
Modify ComponentFactory so that it always takes an engine as an
argument, and update the code that uses it to create one to pass
in as necessary.

Delete some old code for overriding the database URL in the
settings, file, which was for a test that no longer exists (the
relevant code has moved into Safir).

Use the now-session-scoped event loop to run asyncio tasks in the
CLI tests rather than using asyncio.run, which ensures they run in
the same loop in which the engine was created.