Fix excessive middleware stacking during tests
Gafaelfawr was adding Starlette middleware during app startup since
one of the middlewares relies on configuration settings that aren't
available until the configuration is parsed.  However, the middleware
was never removed and the app is a global variable that lives for
the lifetime of the tests, so we were adding an additional copy of
the middleware per test.  When running the whole test suite, this
caused huge amounts of slowness.

Fix this by clearing and rebuilding the middleware stack after each
test.  This is a workaround; a better approach would be to
statically configure the middleware, but it's tricky to see the
best way to do that due to FastAPI's eager object creation.