Remove separate verifier config
Move the verifier configuration into the OIDC provider configuration
because that's the only component that needs to use it.  Stop making
the claims of JWT created by the internal OpenID Connect issuer
depend on the username and UID claims configured for the upstream
provider and instead always use the same claim names.  Tease apart
some internal confusion between issuer configuration and OIDC
provider configuration so that everything is more cleanly separated.