Restructure test JWT code
Move the code for creating test JWTs into a separate module from
the code for making test Gafaelfawr tokens to preserve the internal
distinction between token types.  Remove a now-unnecessary separate
function since the only JWTs we mint are upstream ones.  Remove the
need for that function to be async now that there's a sync way of
getting the configuration.