Support retrieving username from LDAP
For integration with CILogon and COmanage, we'll need to take the
CILogon identifier and use LDAP to resolve it into a username from
COmanage.  Add support for this when using an OpenID Connect
provider.

Add support for simple binds to the LDAP server, required to use
the COmanage LDAP server.