Support UID/GID assignment with Google Firestore
Add support for configuring Firestore (using workload identity for
authentication and authorization) as a repository of UID and GID
assignments.  When this option is enabled, Gafaelfawr will ignore
UID and GID information from the upstream OpenID Connect ID tokens
or from LDAP and instead assign UIDs and GIDs on first use of that
username or group name, or look up the UID and GID from Firestore
if one was already assigned to that username or group name.

To implement this support, retrieval of user information was
extracted from the OpenID Connect authentication provider into a
new internal service, and the cache implementation previously used
only for internal and notebook tokens was made more generic and also
used to cache UIDs and GIDs.  UID and GID information is cached for
the lifetime of the process, since UIDs and GIDs are never reused
and should never change.

UID and GID assignment is done from the ranges documented in
DMTN-225, including assigning UIDs to bot users from a different
range than the UIDs for regular users.