Add enrollment URL support for OpenID Connect
When using CILogon with COmanage, CILogon will pass an opaque
identifier in the sub claim that we need to look up in COmanage
LDAP to resolve to a user.  That was previously implemented, but
with a generic login error if the lookup failed.  However, a
lookup failure is expected for a new user, and rather than an
error, we want to direct that user to the self-service enrollment
flow so that they can create a pending COmanage record and be
added to the queue for approval.

Implement this via a new configuration parameter and a new
distinguished exception raised from the LDAP service layer if
all the queries succeeded but a user record was not found.