Add logging for Firestore
Log whenever changes are made to Firestore, for both initialization
and when a new UID or GID is allocated.  Also log an error if we're
out of UID or GID space or if Firestore is not initialized.

This may log duplicate messages if a transaction is retried, which
seems like acceptable noise.