Move scope determination into user info service
Rather than using a separate function in the login handler, move
determination of token scopes into the user info service where it
fits better.  Use this to retire the get_group_names_from_ldap
method on the user info service, which is no longer required.

This requires a bit of refactoring since the user info service
now needs the top-level Config object.  Preserve a check in the
constructor of the OpenID Connect user info specialization so
that there's no need for asserts or checks for an OIDC configuration
on each call.