Refactor caching layer
Use separate caches for each type of data stored, instead of using
a combined cache for UIDs and GIDs and another combined cache for
internal and notebook tokens.

Replace the dataclass cache objects with real objects that expose
an API and manage their locking.  Move the per-user locking logic
mostly into the cache class itself, and move the cache keys for
internal and notebook tokens into the cache class.

Always take a lock before looking up an internal or notebook token
in the cache, since it seems to make no performance difference and
the code is simpler.