Refactor how per-process context is handled
Gafaelfawr was accumulating an ever-growing list of dependencies,
all of which were called by context_dependency to create a
per-request context.  That context also held all sorts of global
state such as caches which were only used by the component
factory and didn't need to be visible to handlers.

Refactor this by maintaining a single process-wide ProcessContext
that holds the Redis and LDAP connection pools, the caches, and
some similar things that don't vary by request.  Construct that
explicitly from a configuration rather than having some things use
config_dependency on the fly and others not.  Wire together the
necessary mechanisms to regenerate the process context with a new
configuration for the test suite (this won't be used normally during
regular execution of the application).

Rename ComponentFactory to Factory to be shorter and less redundant.