Allow for None request.client
With the latest FastAPI and Starlette, request.client may be None.
Check for this case and extract the client IP in the request
context so that handlers can use it unconditionally.

This takes the approach of throwing an HTTP error if we have no
client IP address available, since we use it for logging and
authentication history.