Fix finding OpenID Connect metadata
The implementation of OpenID Connect metadata discovery, used to
verify an upstream OpenID Connect ID token, did not correctly
follow the OpenID Connect Discovery protocol.  In particular, it
removed the last path component of the issuer URL before appending
the well-known address due to the default behavior of urljoin.

Stop using urljoin and correctly implement the protocol for
forming the URL.