Refactor LDAP testing
Chagne the mock LDAP layer so that the test registers all of the
data instead of returning fixed data.  Parse the search and allow
searching of arbitrary DNs and lists of attributes.

Move the OpenID Connect LDAP tests into a separate file to keep
the number of tests in each file low.