Don't require user metadata in admin token creation
Standardize on semantics where the data stored with the token is
used if present, and if not, data is retrieved from LDAP if
configured.  Avoid storing data derived from LDAP with the token,
and do not store data from upstream OIDC claims if LDAP is
configured and has the corresponding data.

UID is an exception, since a token's UID is fixed at creation,
so look up the user's UID if needed during admin token creation.