Allow users not found in LDAP
If configured to get UIDs from LDAP, users who were not found in
LDAP and didn't have a UID set would result in 500 errors when
requesting user information.  Instead, treat them the same as
users without that data when LDAP was not configured: return the
UID as null.  If the protected application requires a UID, it can
reject the request.  (UIDs are often unnecessary for service to
service authentication.)