Add command to delete all data
Sometimes we'll do destructive upgrades where everyone's username
may change, and in that case we want to purge all Gafaelfawr data
since any remaining data will be meaningless.  Add a command-line
tool to do that.

Drop a workaround in CLI testing from when we were recreating
tables, which is no longer needed now that we're truncating tables
instead.