Generate API documentation with Redoc
Use sphinxcontrib-redoc to generate REST API documentation at
api.html.  Move the internals Python API documentation to
arch/internals, since only developers will need to look at it.
In support of this, add a CLI command to spit out the OpenAPI
schema so that it can be used in the tox build.

I experimented with sphinxcontrib-openapi, but it appears to not
be maintained and has bitrotted, and I couldn't get it to work
with the OpenAPI schema generated by FastAPI.  Redoc does,
provided that I override the version of Redoc with the current
upstream release.

There is as yet no back link from the API documentation to the
rest of the documentation.

Now that the API documentation is of the REST API, move it up in
the documentation index.