Test to ensure Gafaelfawr doesn't accept forms
The web security model for the token API depends on POST requiring
a CORS preflight check from the browser.  This in turn means that it
cannot safely accept form data, since POST of form data doesn't
require a preflight check.  FastAPI supports this inherently, but
add an explciit check to ensure that it doesn't break in some later
version.