Fix editing of token scopes
When a user token was edited to change its scope, but not its
expiration time, its scopes were not updated in Redis.  Since Redis
is canonical for token scopes, this meant that the change appeared
to go through but had no actual effect.  Fixed by updating Redis if
either the scope or expiration of a user token is changed.