Add Slack alerts for uncaught exceptions
Use a wrapped route handler class to catch uncaught exceptions and
report them to a Slack webhook if so configured.

This change is primarily a proof-of-concept for Slack alerting,
which will be used in the future for reporting problems found by
periodic data audits and may be used to report particularly serious
errors that may indicate a broader system problem.  As of this
change, it only handles uncaught exceptions (and tests that none of
the normal errors we are already handling produce Slack messages).