Add support for a primary GID
Add a GID field to the user information.  For GitHub, populate that
with the UID.  For other authentication providers, either retrieve
it from LDAP or populate it with the GID of the user private group
if user private groups are being synthesized.  The GID field can be
set by admins when creating a token directly.