Reformat CHANGELOG.md yet again
Alas the only way to support cutting and pasting the CHANGELOG.md
entry directly into the release notes on GitHub is to use one line
per bullet point, neither wrapped nor with a newline after each
sentence.  Otherwise, GitHub treats the whitespace as significant
and messes up the formatting.  Do that reformatting for all entries
for consistency.

Add internal headings for types of changes to break up the size of
each changelog entry and make it a bit easier to read.