Drop support for /oauth2/callback
We used to have a second name for the /login route to support NCSA
deployments that were registered with CILogon with an older URL
pattern, and a configuration option to tell Gafaelfawr to send that
as the return URL instead.  Those deployments have all been retired,
so we can drop support for this.