Cache empty LDAP group results
The check for cache data for LDAP group results was incorrectly
treating an empty list as a missing cache entry.  As a result, we
were repeatedly doing LDAP searches for group information for bot
users such as mobu that have scopes but no groups.  Check explicitly
for None to determine whether we had a cache hit.