Add additional status logging to commands
Add more logging at the debug level to the non-default commands,
other than the ones that only produce output.  Ensure that each
command logs when it starts and when it is finished, and log before
taking major actions.

For the delete-all-data command, log each action taken at the info
level.