Remove unused X-Auth-Request headers
The original jwt_authorizer software was fairly aggressive about
exposing everything it knew via X-Auth-Request-* headers, and that
behavior has been preserved up until now.  However, most of those
headers were never used, and supporting them is additional unneeded
complexity now that there is an API with user-info and token-info
routes in the rare cases that information is needed.  In some cases,
it shouldn't even be exposed; there is no reaosn for a service to
know all of the scopes a user's token has if they passed an
authorization check.

Drop support for all headers except X-Auth-Request-User,
X-Auth-Request-Email, and X-Auth-Request-Token, which are in active
use.  Adjust the test suite accordingly, adding calls to user-info
and token-info where appropriate to keep applying the same checks.