Don't allow user token editing
Allowing a user to change the expiration and scopes of a token after
it's been created is not a common operation to support in other
implementations, such as GitHub, and may be confusing or have
unforseen security issues.  On the other hand, we have tested
and debugged code to ensure it works.

Compromise by retaining the tested functionality but limit it to
only token administrators.  Remove the UI for a user to edit their
own tokens, and instead expect them to delete the token and create
a new one if they need to change something.