Rewrite documentation using new Documenteer
Rework the documentation to use the new Documenteer system and
theme.  Use this as an opportunity to restructure the documentation
in a way that will hopefully be more logical and oriented towards
the reader, rather than when I happened to think of something.

Use sphinx-diagrams to generate the architecture diagram instead
of generating it outside of Sphinx.

Create the directory where the OpenAPI schema is written if it didn't
already exist.