Set a default minimum lifetime
If and only if a delegated token is requested, force the minimum
lifetime to be at least MINIMUM_LIFETIME (five minutes).  This will
hopefully fix a problem that caused an exception in production due
to an invalid Redis key expiration time, presumably because someone
attempted to authenticate with a token that was about to expire,
resulting in a delegated token with a lifetime of 0 seconds.