Convert Kubernetes operator to Kopf
In preparation for adding more Kubernetes operator functionality,
convert the Kubernetes operator to Kopf.  The Kopf framework now
handles monitoring Kubernetes for changes, deduplicating events,
and storing status in Kubernetes, which simplifies the internal
logic (at the cost of having somewhat less control over how status
is stored).

Unfortunately, Kopf does not support testing against a mock
Kubernetes, so testing is now done only via a new tox py-full
target that enables Kubernetes tests using the current default
Kubernetes environment.  In CI, this is done using Minikube.
These tests involve a fair number of somewhat arbitrary delays to
allow the operator to work in the background, which is the
recommended testing strategy in the Kopf documentation.

For now at least, remove the tests for specific logging from the
Kubernetes operator, since integrating that with Kopf and ignoring
Kopf's own logs is tricky and the value of the tests is not that
high.