Support specifying GitHub org/team pairs for scopes
Allow the values of the group_mapping dict to be a list of either
strings as before (still group names) or a dict with one key,
github, whose value is a dict with keys organization and team.
(This structure is to allow for future extensibility if needed.)
This specifies a GitHub team and Gafaelfawr will transform this
into a group name internally.