Create namespace for each Kubernetes test
Rather than trying to reuse the same namespaces, create a new
Kubernetes namespace for each test.  This allows us to delete it
but not care when the deletion finally happens, since the next test
will use a new namespace.

This drops the testing across multiple namespaces in one run of the
operator, but this doesn't seem like a major loss since we're still
testing that the operator looks at every namespace.