Add a schema for GafaelfawrServiceToken status
I had deleted the schema for the GafaelfawrServiceToken status
section, since Kopf also uses it and the format had changed.
However, the create and periodic portions of the status do have
a reliable schema.  Only the Kopf stuff may change independently
and is handled by the x-kubernetes-preserve-unknown-fields
annotation.

Also update additionalPrinterColumns to not refer to status
columns that had been removed.