Add documentation for GafaelfawrIngress
Document GafaelfawrIngress as the default way to configure an
ingress.  Add a bit more explicit documentation of how Gafaelfawr
works.  Keep the documentation for how to configure it with
annotations, since sometimes this is required for third-party
charts, but move the bulk of the details and discussion to the
GafaelfawrIngress documentation.