Refactor GafaelfawrIngress validation
Separate the GafaelfawrIngressScopes models into two with a shared
base class to capture all of the validation logic in the type system.
Do the same for the service port.

I considered also doing the same for the delegated tokens, but the
right abstract properties are less obvious there, so keep the current
structure and use a new validation function that ensures that only
one of a group of properties are set.