Restructure the ingress documentation
Break out the overview of how Gafaelfawr ingress protection works
into a separate page so that it can be linked from both the
GafaelefawrIngress page and the manual ingress page, and rename
the former accordingly.  Move prerequisites into the single
prerequisites page instead of having a second copy in the ingress
page.  Reword and restructure things a bit.