Change some Optional to | None
Take advantage of the new syntax to more explicitly state when
None is a valid value.  Keep using Optional only for cases where
the whole parameter to the function or attribute in the model is
optional, or when None is used only as a temporary placeholder for
state that has to be initialized.