Invalidate the LDAP cache if login was rejected
If we reject the login because the user is not a member of any
valid groups, invalidate the LDAP cache so that the next attempt
will check LDAP again.  The user is probably going to try to fix
the problem right away, and we don't want to make them wait for the
LDAP cache to time out before they can see if their fix worked.