Add Authorization and Cookie header filtering
Reflect the Authorization and Cookie headers from the request back
in the response from the /auth route with Gafaelfawr tokens and
cookies filtered out.  Lift those headers into the request to the
protected service via NGINX Ingress annotation in the Ingress
resources created from GafaelfawrIngress, and document how to do
this for manually-configured ingresses.

This prevents leaking the user's original credentials to protected
services, which could reuse them to impersonate the user beyond the
scope of their delegated tokens.