Support passing delegated token in Authorization
Add new config.delegate.useAuthorization GafaelfawrIngress setting
and use_authorization query parameter to the /auth route that also
puts a copy of any delegated token into the Authorization header as
a bearer token.  The CADC TAP service requires the token be passed
that way to work well with their authentication layer, and this
removes the need for a custom NGINX configuration snippet.