Add support for anonymous ingresses
Allow a GafaelfawrIngress to specify that it is anonymous.  In this
case, most of the configuration is not allowed, but Gafaelfawr is
still configured as the auth-uri so that it can strip tokens and
cookies from the Authorization and Cookie headers.  Add a new route,
/auth/anonymous, that does only this stripping.

Also diagnose a GafaelfawrIngress that requests a basic authentication
challenge but also wants an auth redirect, since those are mutually
incompatible.