Accept tokens in either field of Basic Auth
Rather than requiring the user use an x-oath-basic placeholder in
the other field, take advantage of the fact that our tokens have
a well-known format and accept tokens in either the username or the
password field.  If both fields are tokens, they must match.

The original design was based on an earlier iteration of how GitHub
handled Basic Auth, but GitHub now supports putting the token in
either field without any special marking, which seems more convenient
and easier to explain.