Support camel-case in configuration
To avoid the divergence between the on-disk configuration file and
the Helm values file in Phalanx, enable camel-case support for the
parsing of the disk configuration file.  Update all examples and
test configuration files accordingly.

Stop using BaseSettings because we don't want to support environment
variables or .env files, and those are the only things BaseSettings
adds over BaseModel.

Enable slots for the frozen configuration for minor performance and
memory benefits.