Filter out some deprecation warnings
The google.* modules are using the deprecated pkg_resources API,
and Kopf is doing soemthing deprecated with iterators. We can't
do anything about either of these problems, so suppress the
warnings so that they don't add clutter to the pytest output.

Also add the comment explaining why we add to python_files in the
pytest configuration.