Add a data model for Slack messages
Rather than doing all of the formatting directly in the Slack client,
add a data model for Slack messages that allows representation of
structured messages and hopefully will also be sufficient to handle
the needs of mobu. Use that model to do field truncation to match the
Slack API limits.

Move the uncaught exception to an attachment for consistency with how
mobu formats messages, which will make it easier to share the Slack
client.