Improve data available during error handling
Use a wrapper exception around httpx's HTTPError so that it can be
ready for reporting to Slack, and add new functionality to the
SlackClient (not yet used) to report a caught exception.

Convert the server-side exceptions for user logins and retrieving
user metadata to SlackExceptions so that they can be reported to
Slack. Add the user as metadata in the exception where it is known.

Add the user in more places in the logging metadata when logging
messages. Previously, many messages were logged without a user if
the failures were too low in the stack or happened while the user's
information was being constructed. Now, we should have user in the
metadata wherever possible.

Move a bit more information out of the log message into structured
data.