Refactor Slack uncaught exception reporting again
After further consideration, get rid of the Slack client dependency
and consolidate all the Slack alert reporting code into a single
module. Turn the global shared alert client for routers into a
class variable on SlackRouteErrorHandler and move the initialize
method that was handled by the dependency to a class method. This is
equivalent to a global variable (which seems unavoidable without too
much complexity), but keeps things grouped in a way that is hopefully
easier to understand.

Move creation of a Slack client for reporting other errors or
messages to the factory, like any other object that depends on
Gafaelfawr configuration.