Report user-info route errors to Slack
If Gafaelfawr fails to retrieve information about a user from an
external user information service, catch that exception and report
it properly to slack rather than leaving it uncaught. This is a
somewhat-expected error in the sense that we know external services
will sometimes go down, so we shouldn't leave it uncaught.