Use full paths in handlers
Rather than loading the API handler with a path prefix, use the
full paths when registering routes in each of the routers. This
makes the code a bit less confusing.

Other Safir-based applications also support a path_prefix
configuration option, but Gafaelfawr takes over various top-level
paths in the environment (such as /login, /logout, and
/.well-known paths), so this doesn't make sense for it.