Report token deserialization errors to Slack
If we cannot deserialize the token data from Redis, report that
to Slack since that indicates some sort of infrastructure failure.

Lower the logging about an incorrect secret portion of a token to
a warning from an error, since this is really just a special case
of having an invalid token (if unlikely outside of an attempted
compromise).