Redo Redis connection configuration
Since adding more connection retry configuration, we have been seeing
periodic "Connection closed by server" errors in all environments.
One hypothesis is that setting socket_keepalive is keeping the
connection open longer than the Redis server permits, and it's changing
the error in a way that's causing this problem. Try dropping that
configuration.

Stop setting an explicit list of exceptions on which to retry, since
the default seems to be fine.