Hopefully fix Redis connection pooling
Gafaelfawr was incorrectly using the sync version of Retry. Use the
asyncio one instead so that we call the correct sleep.

Create the connection pool explicitly so that we can cap its size
(by default, Redis has an unbounded connection pool size) and use
the BlockingConnectionPool variant so that requests wait for a
connection to become available.