Switch minikube setup actions
The GitHub Action to set up minikube that the test suite was using
is apparently no longer supported, so switch to the one that's
recommended by the minikube documentation and is now used in Phalanx.

Remove some unnecessary name settings for action steps, and rename
the periodic workflow file to periodic-ci, matching the current
templates.