Add better error reporting of linkcheck failures
Sphinx's linkcheck target does not summarize errors at the end of
the run normally, but it does write the errors into an output.txt
file. Move the linkcheck code into Makefile so that we can use
shell trickery and display output.txt at the end of the run if the
check failed.

Add a make help target as the default that displays the available
Makefile targets.