Convert to Pydantic v2
Update to Pydantic v2 and remove all calls to deprecated methods.
Use alpha releases of Safir and documenteer.

This unfortunately requires two ugly workarounds, one due to a bug
in Pydantic's mypy plugin and one due to a bug in FastAPI. For the
first, a bunch of fields in TokenData that should be inherited from
TokenUserInfo have to be reiterated because Pydantic's mypy plugin
loses a union with None on inherited attributes. For the second, an
exception handler for RequestValidationError had to be added to
avoid FastAPI attempting to serialize PydanticUndefinedType when the
user doesn't provide a query attribute that must be a list.